requirements.txt
discord.py
pymongo
keyboard